+++ http/src/request/application/interaction/create_response.rs
@@ -1,25 +1,25 @@
 use crate::{
     client::Client,
-    error::Error,
-    request::{Request, TryIntoRequest},
+    error::Error as HttpError,
+    request::{attachment::AttachmentManager, Request, TryIntoRequest},
     response::{marker::EmptyBody, ResponseFuture},
     routing::Route,
 };
 use twilight_model::{
-    application::callback::InteractionResponse,
+    http::interaction::InteractionResponse,
     id::{marker::InteractionMarker, Id},
 };
 
-/// Respond to an interaction, by ID and token.
+/// Respond to an interaction, by its ID and token.
 #[must_use = "requests must be configured and executed"]
-pub struct InteractionCallback<'a> {
+pub struct CreateResponse<'a> {
     interaction_id: Id<InteractionMarker>,
     interaction_token: &'a str,
     response: &'a InteractionResponse,
     http: &'a Client,
 }
 
-impl<'a> InteractionCallback<'a> {
+impl<'a> CreateResponse<'a> {
     pub(crate) const fn new(
         http: &'a Client,
         interaction_id: Id<InteractionMarker>,
@@ -47,17 +47,37 @@ impl<'a> InteractionCallback<'a> {
     }
 }
 
-impl TryIntoRequest for InteractionCallback<'_> {
-    fn try_into_request(self) -> Result<Request, Error> {
-        let request = Request::builder(&Route::InteractionCallback {
+impl TryIntoRequest for CreateResponse<'_> {
+    fn try_into_request(self) -> Result<Request, HttpError> {
+        let mut request = Request::builder(&Route::InteractionCallback {
             interaction_id: self.interaction_id.get(),
             interaction_token: self.interaction_token,
-        })
-        .json(self.response)?
-        .use_authorization_token(false)
-        .build();
+        });
 
-        Ok(request)
+        // Interaction executions don't need the authorization token, only the
+        // interaction token.
+        request = request.use_authorization_token(false);
+
+        // Determine whether we need to use a multipart/form-data body or a JSON
+        // body.
+        if let Some(attachments) = self
+            .response
+            .data
+            .as_ref()
+            .and_then(|data| data.attachments.as_ref())
+        {
+            let fields = crate::json::to_vec(&self.response).map_err(HttpError::json)?;
+
+            let form = AttachmentManager::new()
+                .set_files(attachments.iter().collect())
+                .build_form(&fields);
+
+            request = request.form(form);
+        } else {
+            request = request.json(&self.response)?;
+        }
+
+        Ok(request.build())
     }
 }
 
@@ -66,7 +86,10 @@ mod tests {
     use crate::{client::Client, request::TryIntoRequest};
     use std::error::Error;
     use twilight_http_ratelimiting::Path;
-    use twilight_model::{application::callback::InteractionResponse, id::Id};
+    use twilight_model::{
+        http::interaction::{InteractionResponse, InteractionResponseType},
+        id::Id,
+    };
 
     #[test]
     fn test_interaction_callback() -> Result<(), Box<dyn Error>> {
@@ -76,10 +99,14 @@ mod tests {
 
         let client = Client::new(String::new());
 
-        let sent_response = InteractionResponse::DeferredUpdateMessage;
+        let response = InteractionResponse {
+            kind: InteractionResponseType::DeferredUpdateMessage,
+            data: None,
+        };
+
         let req = client
             .interaction(application_id)
-            .interaction_callback(interaction_id, &token, &sent_response)
+            .create_response(interaction_id, &token, &response)
             .try_into_request()?;
 
         assert!(!req.use_authorization_token());